postgresql-vector/requirements.txt
python-dotenv
psycopg[binary]
sqlalchemy
pgvector
openai

spacy

boto3
botocore
awscli